+++ src/resdk/tables/__init__.py
@@ -44,6 +44,7 @@ Table classes
 from .methylation import MethylationTables  # noqa
 from .microarray import MATables  # noqa
 from .ml_ready import MLTables  # noqa
+from .qc import QCTables  # noqa
 from .rna import RNATables  # noqa
 from .variant import VariantTables  # noqa
 
@@ -51,6 +52,7 @@ __all__ = (
     "MATables",
     "MLTables",
     "MethylationTables",
+    "QCTables",
     "RNATables",
     "VariantTables",
 )