+++ tests/unit/test_tables.py
@@ -239,11 +239,6 @@ class TestTables(unittest.TestCase):
         self.assertTrue(time() - t < 0.1)
         self.assertIs(mapping, self.gene_map)
 
-    @patch("resdk.tables.base.clear_cache_dir_resdk")
-    def test_clear_cache(self, clear_mock):
-        RNATables.clear_cache()
-        clear_mock.assert_called()
-
     def test_metadata_version(self):
         self.collection.samples.get = self.web_request(self.sample)
         self.collection.relations.get = self.web_request(self.relation)
@@ -263,23 +258,6 @@ class TestTables(unittest.TestCase):
         ct1 = RNATables(self.collection)
         with self.assertRaises(ValueError):
             version = ct1._metadata_version
-
-    def test_qc_version(self):
-        self.collection.data.filter().iterate = self.web_request([self.data])
-
-        ct = RNATables(self.collection)
-        version = ct._qc_version
-        self.assertEqual(version, str(hash(tuple([12345]))))
-
-        # use cache
-        t = time()
-        version = ct._qc_version
-        self.assertTrue(time() - t < 0.1)
-
-        self.collection.data.filter().iterate = self.web_request([])
-        ct1 = RNATables(self.collection)
-        with self.assertRaises(ValueError):
-            version = ct1._qc_version
 
     def test_data_version(self):
         ct = RNATables(self.collection)